requirements.txt

fastapi
uvicorn
jinja2
python-multipart
nltk
requests
python-dotenv
anthropic